List of ip numbers running Mastodon that try to connect even when ActivityPub is disabled
I tried to reason with the developers of Mastodon but they have a weird way to solving how to unlist from being visited. They do not listen to the regular server status messages like a 404 or 403. They will keep trying forever. 

I would love to enable AP again, had only 5 connections on it, but they want you to generate a 401, a 429 or a 5xx. That results them pounding my server with a request every minute for seven days. A pure waste of bandwidth and energy imho.